mcp/adapter/smartui_mcp/requirements.txt
# SmartUI MCP - 智慧感知UI服务

# 核心依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# 数据处理
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0

# 模板和UI
jinja2>=3.1.0
markupsafe>=2.1.0
pillow>=10.1.0

# 配置和日志
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=23.2.0
colorama>=0.4.6

# 异步和并发
asyncio-mqtt>=0.16.0
asyncio-throttle>=1.0.2
aioredis>=2.0.0

# 机器学习和AI
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.1.0

# 文件监控
watchdog>=3.0.0

# HTTP客户端
httpx>=0.25.0
requests>=2.31.0

# 时间处理
python-dateutil>=2.8.0
pytz>=2023.3

# 加密和安全
cryptography>=41.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# 测试依赖（开发环境）
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # 用于测试HTTP客户端

# 开发工具
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# 文档生成
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# 性能监控
psutil>=5.9.0
memory-profiler>=0.61.0

# 数据验证
jsonschema>=4.20.0
cerberus>=1.3.5

# 网络工具
dnspython>=2.4.0
netifaces>=0.11.0

